+++ src/showdown-katex.js
@@ -1,13 +1,14 @@
-/** !
- * @file showdown-katex: markdown + ( latex and/or asciimath )
- * @author obedm503
- * @git https://github.com/obedm503/showdown-katex.git
- * @examples https://obedm503.github.io/showdown-katex/
- * @license MIT
- */
 import katex from 'katex';
 import renderMathInElement from 'katex/dist/contrib/auto-render';
+import showdown from 'showdown';
 import asciimathToTex from './asciimath-to-tex';
+
+if (process.env.TARGET === 'cjs') {
+  const { JSDOM } = require('jsdom');
+  const jsdom = new JSDOM();
+  global.DOMParser = jsdom.window.DOMParser;
+  global.document = jsdom.window.document;
+}
 
 /**
  * @param {object} opts
@@ -111,8 +112,6 @@ const showdownKatex = userConfig => () => {
 };
 
 // register extension with default config
-if (typeof window.showdown !== 'undefined') {
-  window.showdown.extension('showdown-katex', showdownKatex());
-}
+showdown.extension('showdown-katex', showdownKatex());
 
 export default showdownKatex;